+++ project/v2/shop/app/shop_srv/goods/srv/internal/data/v1/goods.go
@@ -10,7 +10,7 @@ import (
 
 type GoodsStore interface {
 	Get(ctx context.Context, ID uint64) (*do.GoodsDO, error)
-	ListByIDs(ctx context.Context, ids []uint32, orderby []string) (*do.GoodsDOList, error)
+	ListByIDs(ctx context.Context, ids []uint64, orderby []string) (*do.GoodsDOList, error)
 	List(ctx context.Context, opts metav1.ListMeta, orderby []string) (*do.GoodsDOList, error)
 	Create(ctx context.Context, goods *do.GoodsDO) error
 	CreateInTxn(ctx context.Context, txn *gorm.DB, goods *do.GoodsDO) error
